+++ wifipumpkin3/core/common/console.py
@@ -3,7 +3,7 @@ from wifipumpkin3.core.controllers.defaultcontroller import *
 from wifipumpkin3.core.config.globalimport import *
 
 from wifipumpkin3.modules import *
-from wifipumpkin3.modules import module_list, all_modules
+from wifipumpkin3.modules import module_list
 
 # This file is part of the wifipumpkin3 Open Source Project.
 # wifipumpkin3 is licensed under the Apache 2.0.
@@ -314,10 +314,6 @@ class PumpkinShell(Qt.QObject, ConsoleUI):
         if logger != None:
             return logger.setIgnore(False)
         print(display_messages("Logger class not found.", error=True))
-
-    def do_clients(self, args):
-        """ap: show all connected clients on AP"""
-        self.uiwid_controller.ui_table_mod.start()
 
     def do_stop(self, args):
         """ap: stop access point service"""