+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added 
 - added: param -iNM or -ignore-from-networkmanager for ignore interface from network-manager #158
+- fixed: Emoji crashes scans #104
 
 ## [Released]